Show pretty start/due dates on the dashboard
