+++ orchestra/static/orchestra/timing/datetime-display/datetime-display.directive.es6.js
@@ -1,0 +1,35 @@
+import template from './datetime-display.html'
+import moment from 'moment-timezone'
+
+export default function datetimeDisplay () {
+  return {
+    template,
+    restrict: 'E',
+    scope: {
+      datetime: '=',
+      showTime: '=?'
+    },
+    link: (scope, elem, attrs) => {
+      if (!scope.showTime) {
+        scope.showTime = true
+      }
+      const getLocalTime = (datetimeString) => {
+        return datetimeString ? moment.utc(datetimeString).tz(moment.tz.guess()) : null
+      }
+
+      const getPrettyDatetime = (datetime) => {
+        if (scope.showTime) {
+          return datetime.format('ddd, MMM D h:mm a')
+        } else {
+          return datetime.format('ddd, MMM D')
+        }
+      }
+
+      scope.getDatetimeDisplay = (datetimeString) => {
+        const datetime = getLocalTime(datetimeString)
+        const datetimeInfo = datetime ? `${getPrettyDatetime(datetime)}` : ''
+        return datetimeInfo
+      }
+    }
+  }
+}
