+++ orchestra/static/orchestra/timing/datetime-display/datetime-display.html
@@ -1,0 +1,1 @@
+<span>{{getDatetimeDisplay(datetime)}}</span>
