+++ orchestra/static/orchestra/timing/timing.module.es6.js
@@ -10,6 +10,7 @@ import workTimer from 'orchestra/timing/timer/timer.directive.es6.js'
 import timeEntries from 'orchestra/timing/time-entries.service.es6.js'
 import TimeEntry from 'orchestra/timing/time-entry.service.es6.js'
 import timeInput from 'orchestra/timing/time-input/time-input.directive.es6.js'
+import datetimeDisplay from 'orchestra/timing/datetime-display/datetime-display.directive.es6.js'
 
 const name = 'orchestra.timing'
 angular.module(name, ['ui.select', 'ngSanitize', common])
@@ -18,6 +19,7 @@ angular.module(name, ['ui.select', 'ngSanitize', common])
   .directive('enforceIntegers', enforceIntegers)
   .directive('workTimer', workTimer)
   .directive('timeInput', timeInput)
+  .directive('datetimeDisplay', datetimeDisplay)
   .controller('TimecardController', TimecardController)
   .factory('timeEntries', timeEntries)
   .factory('TimeEntry', TimeEntry)
