+++ orchestra/static/orchestra/todos/todo-checklist.directive.es6.js
@@ -1,5 +1,4 @@
 import template from './todo-checklist.html'
-import moment from 'moment-timezone'
 import './todo-checklist.scss'
 
 export default function todoChecklist () {
@@ -14,21 +13,7 @@ export default function todoChecklist () {
       steps: '<',
       taskSlugs: '<'
     },
-    link: function (scope, elem, attrs) {
-      const getLocalTime = (datetimeString) => {
-        return datetimeString ? moment.utc(datetimeString).tz(moment.tz.guess()) : null
-      }
-
-      const getPrettyDatetime = (datetime) => {
-        return datetime.format('ddd, MMM D h:mm a')
-      }
-
-      scope.getDatetimeDisplay = (datetimeString) => {
-        const datetime = getLocalTime(datetimeString)
-        const datetimeInfo = datetime ? `${getPrettyDatetime(datetime)}` : ''
-        return datetimeInfo
-      }
-
+    link: (scope, elem, attrs) => {
       scope.isNonEmptyString = (str) => {
         return str !== null && str !== undefined && str !== ''
       }
