+++ orchestra/static/orchestra/todos/todo-checklist.html
@@ -10,11 +10,11 @@
       <span class="todo__description">{{todo.description}}</span>
       <span class="todo__dates">
         <span ng-if="isNonEmptyString(todo.start_by_datetime)">
-          Start {{getDatetimeDisplay(todo.start_by_datetime)}}
+          Start <datetime-display datetime="todo.start_by_datetime" />
         </span>
         <span ng-if="isNonEmptyString(todo.start_by_datetime) && isNonEmptyString(todo.due_datetime)" class="todo__dates-separator">|</span>
         <span ng-if="isNonEmptyString(todo.due_datetime)">
-          Due {{getDatetimeDisplay(todo.due_datetime)}}</span>
+          Due <datetime-display datetime="todo.due_datetime" /></span>
       </span>
     </label>
   </div>
